moves search with filters to search folder
