+++ frontend/components/products/product_index_container.jsx
@@ -1,7 +1,8 @@
 import { connect } from 'react-redux';
 import ProductIndex from './product_index';
 
-const msp = (products) => {
+const msp = (state) => {
+   const products = state.entities.products;
    return {
       products
    };
