+++ frontend/components/search/search_page.jsx
@@ -1,15 +1,10 @@
 import React from 'react';
 import ProductIndexContainer from '../products/product_index_container';
-import SearchBarContainer from '../search/search_bar_container';
+import SearchWithFiltersContainer from './search_with_filters_container';
 
-class Home extends React.Component {
+class SearchPage extends React.Component {
     constructor(props){
         super(props);
-        this.products = this.props.products;
-        this.state = {
-            category: ""
-        }
-        this.changeCategory = this.changeCategory.bind(this);
     }
 
     changeCategory(e) {
@@ -21,17 +16,12 @@ class Home extends React.Component {
     render(){
         return (
             <>
-                I'm home!
-                <SearchBarContainer />
-                <button onClick={this.changeCategory} value={'surf'}>Surf</button>
-                <button onClick={this.changeCategory} value={'skate'}>Skate</button>
-                <button onClick={this.changeCategory} value={'snow'}>Snow</button>
-                <br />
-                <button>Advanced</button>
-                <ProductIndexContainer products={ this.products } />
+                <strong>SEARCH!?!?!?!?!!?!?!!!!!!</strong>
+                <SearchWithFiltersContainer />
+                <ProductIndexContainer />
             </>
         )
     }
 }
 
-export default Home; 
+export default SearchPage;
