+++ frontend/components/search/search_with_filters_container.jsx
